:lipstick: Divider 컴포넌트 Layout Styled 파일에 분리
